mcp_server/requirements.txt
chromadb==1.3.5
fastmcp==2.13.3
pandas==2.3.3
numpy==2.3.5
pillow==12.0.0
open_clip_torch==3.2.0
python-dotenv==1.2.1